[INLONG-10784][Sort] Rename row module structure
