+++ inlong-sort/sort-formats/format-row/format-inlongmsg-base/src/main/java/org/apache/inlong/sort/formats/inlongmsg/row/AbstractInLongMsgMixedFormatDeserializer.java
@@ -15,18 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.sort.formats.inlongmsg;
+package org.apache.inlong.sort.formats.inlongmsg.row;
 
-import org.apache.flink.api.common.functions.FlatMapFunction;
-import org.apache.flink.api.java.typeutils.ResultTypeQueryable;
-import org.apache.flink.types.Row;
+import org.apache.inlong.sort.formats.inlongmsg.FailureHandler;
+
+import javax.annotation.Nonnull;
 
 /**
- * The converter for a mixed inlongmsg format.
+ * The base for all inlongmsg mixed format deserializers.
  */
-public interface InLongMsgMixedFormatConverter
+public abstract class AbstractInLongMsgMixedFormatDeserializer
         extends
-            FlatMapFunction<Row, Row>,
-            ResultTypeQueryable<Row> {
+            AbstractInLongMsgFormatDeserializer {
 
+    @Deprecated
+    public AbstractInLongMsgMixedFormatDeserializer(@Nonnull Boolean ignoreErrors) {
+        this(InLongMsgUtils.getDefaultExceptionHandler(ignoreErrors));
+    }
+
+    public AbstractInLongMsgMixedFormatDeserializer(@Nonnull FailureHandler failureHandler) {
+        super(failureHandler);
+    }
 }
