+++ inlong-sort/sort-formats/format-row/format-inlongmsg-base/src/test/java/org/apache/inlong/sort/formats/inlongmsg/InLongMsgRowDataSerDeTest.java
@@ -33,6 +33,7 @@ import org.apache.flink.table.data.TimestampData;
 import org.apache.flink.table.data.binary.BinaryStringData;
 import org.apache.flink.table.factories.utils.FactoryMocks;
 import org.apache.flink.table.runtime.connector.source.ScanRuntimeProviderContext;
+import org.apache.inlong.sort.formats.inlongmsg.row.InLongMsgFormatFactory;
 import org.junit.Test;
 
 import java.io.IOException;
